Move graphql stuff out of server.js
