+++ server.js
@@ -1,56 +1,13 @@
 'use strict';
 
-const { merge } = require('lodash');
-const { userSchema } = require('./schema/user');
-const { userResolvers } = require('./resolvers/user');
-const { eventSchema } = require('./schema/event');
-const { eventResolvers } = require('./resolvers/event');
-
 const express = require('express');
 const bodyParser = require('body-parser');
 const { graphqlExpress, graphiqlExpress } = require('apollo-server-express');
-const { GraphQLDate, GraphQLTime, GraphQLDateTime } = require('graphql-iso-date');
-const { makeExecutableSchema } = require('graphql-tools');
-
-
-const { Sequelize } = require('sequelize');
+const { schema } = require('./graphql');
 
 // Constants
 const PORT = 8080;
 const HOST = '0.0.0.0';
-
-// Schema
-
-const scalars = `
-  scalar Date
-  scalar Time
-  scalar DateTime
-`;
-
-const Query = `
-  type Query {
-    _empty: String
-  }
-`;
-
-const Mutation = `
-  type Mutation {
-    _empty: String
-  }
-`;
-
-const resolvers = {
-  // Custom scalars
-  Date: GraphQLDate,
-  Time: GraphQLTime,
-  DateTime: GraphQLDateTime,
-
-};
-
-const schema = makeExecutableSchema({
-  typeDefs: [ scalars, Query, Mutation, userSchema, eventSchema ],
-  resolvers: merge(resolvers, userResolvers, eventResolvers),
-});
 
 // App
 const app = express();
